requirements.txt
torch
transformers
pandas
numpy
Pillow
scikit-learn
torchvision
streamlit